requirements.txt
Django==1.8.2
GDAL==1.11.2
Markdown==2.5.2
Pillow==2.7.0
PyJWT==1.3.0
PyYAML==3.11
Pygments==1.6
amqp==1.4.6
anyjson==0.3.3
arrow==0.5.4
billiard==3.3.0.19
boto==2.36.0
celery==3.1.17
django-celery==3.1.16
django-debug-toolbar==1.3.0
django-extensions==1.5.2
django-extra-fields==0.3
django-filter==0.9.1
django-push-notifications==1.2.0
django-rest-swagger==0.3.2
django-suit==0.2.13
django-uuidfield==0.5.0
djangorestframework==3.1.3
docutils==0.12
drf-extensions==0.2.6
enum34==1.0.4
frosted==1.4.1
geopy==1.10.0
ipaddress==1.0.7
itsdangerous==0.24
jsonfield==1.0.3
kombu==3.0.24
mock==1.0.1
nose==1.3.6
numpy==1.9.2
oauthlib==0.7.2
pies==2.6.3
pies2overrides==2.6.3
pilkit==1.1.12
pjson==0.6
psycopg2==2.5.4
pyparsing==2.0.3
python-dateutil==2.4.2
python-openid==2.2.5
python-social-auth==0.2.10
pytz==2015.2
requests==2.7.0
requests-oauthlib==0.5.0
rollbar==0.9.9
scikit-learn==0.16.1
scipy==0.15.1
six==1.9.0
sqlparse==0.1.15
uWSGI==2.0.10
ujson==1.33
wsgiref==0.1.2
xmlformatter==0.1.1
